Rename configuration.ts to index.ts
